backend/requirements.txt
# Backend Requirements for Car Loan Assistant

fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
requests==2.31.0
openai==1.6.1
PyPDF2==3.0.1
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0
pdfminer.six==20231228
python-dotenv==1.0.0
python-docx==1.1.0
